Implement queue insert and remove functions
In the insertion part, each malloc call is checked for success. The
input string is copied into the value, and then the element is added to
the queue.
In the removal part, list_first_entry is used and check if the queue is
empty. After removal, the value is stored in the sp pointer.

Change-Id: I876b8be6657e9e73a31600da1b20fcc646bd1a67